server/requirements.txt
Flask>=2.3.0
whisper @ git+https://github.com/openai/whisper.git
edge-tts>=0.3.0
google-generativeai>=0.1.0